folder restructure for tests and components
