+++ components/nav/navItem.module.scss
@@ -1,4 +1,4 @@
-@use './utils/library';
+@use '../../styles/utils/library';
 
 .navContainer {
   display: flex;
