+++ components/nav/navbar.module.scss
@@ -1,4 +1,4 @@
-@use '../styles/utils/library';
+@use '../../styles/utils/library';
 
 .container {
  padding: 10px;
